+++ src/Modules/Admin/Controllers/HotspotAssetController.php
@@ -39,7 +39,16 @@ class HotspotAssetController extends Controller
                 'rules' => [
                     [
                         'allow' => true,
-                        'actions' => ['create', 'delete', 'duplicate', 'index', 'order', 'status', 'update'],
+                        'actions' => [
+                            'create',
+                            'delete',
+                            'delete-all',
+                            'duplicate',
+                            'index',
+                            'order',
+                            'status',
+                            'update',
+                        ],
                         'roles' => [Entry::AUTH_ENTRY],
                     ],
                 ],
@@ -82,6 +91,11 @@ class HotspotAssetController extends Controller
     public function actionDuplicate(int $id): Response|string
     {
         return $this->duplicateAsset($this->findHotspotAsset($id));
+    }
+
+    public function actionDeleteAll(?int $hotspot = null): Response
+    {
+        return $this->deleteAssets($this->findHotspotModel($hotspot));
     }
 
     public function actionOrder(?int $hotspot = null): string